+++ examples/01-basic/04-all-blocks/App.tsx
@@ -1,36 +1,11 @@
-import {
-  BlockNoteSchema,
-  combineByGroup,
-  filterSuggestionItems,
-  locales,
-  withPageBreak,
-} from "@blocknote/core";
 import "@blocknote/core/fonts/inter.css";
 import { BlockNoteView } from "@blocknote/mantine";
 import "@blocknote/mantine/style.css";
-import {
-  SuggestionMenuController,
-  getDefaultReactSlashMenuItems,
-  useCreateBlockNote,
-  getPageBreakReactSlashMenuItems,
-} from "@blocknote/react";
-import {
-  getMultiColumnSlashMenuItems,
-  multiColumnDropCursor,
-  locales as multiColumnLocales,
-  withMultiColumn,
-} from "@blocknote/xl-multi-column";
-import { useMemo } from "react";
+import { useCreateBlockNote } from "@blocknote/react";
 
 export default function App() {
   // Creates a new editor instance.
   const editor = useCreateBlockNote({
-    schema: withMultiColumn(withPageBreak(BlockNoteSchema.create())),
-    dropCursor: multiColumnDropCursor,
-    dictionary: {
-      ...locales.en,
-      multi_column: multiColumnLocales.en,
-    },
     initialContent: [
       {
         type: "paragraph",
@@ -52,35 +27,6 @@ export default function App() {
       {
         type: "paragraph",
         content: "Paragraph",
-      },
-      {
-        type: "columnList",
-        children: [
-          {
-            type: "column",
-            props: {
-              width: 0.8,
-            },
-            children: [
-              {
-                type: "paragraph",
-                content: "Hello to the left!",
-              },
-            ],
-          },
-          {
-            type: "column",
-            props: {
-              width: 1.2,
-            },
-            children: [
-              {
-                type: "paragraph",
-                content: "Hello to the right!",
-              },
-            ],
-          },
-        ],
       },
       {
         type: "heading",
@@ -120,7 +66,6 @@ export default function App() {
           ],
         },
       },
-      { type: "pageBreak" },
       {
         type: "file",
       },
@@ -192,21 +137,6 @@ export default function App() {
     ],
   });
 
-  const slashMenuItems = useMemo(() => {
-    return combineByGroup(
-      getDefaultReactSlashMenuItems(editor),
-      getPageBreakReactSlashMenuItems(editor),
-      getMultiColumnSlashMenuItems(editor)
-    );
-  }, [editor]);
-
   // Renders the editor instance using a React component.
-  return (
-    <BlockNoteView editor={editor} slashMenu={false}>
-      <SuggestionMenuController
-        triggerCharacter={"/"}
-        getItems={async (query) => filterSuggestionItems(slashMenuItems, query)}
-      />
-    </BlockNoteView>
-  );
+  return <BlockNoteView editor={editor} />;
 }
