+++ packages/core/src/editor/Block.css
@@ -311,15 +311,8 @@ NESTED BLOCKS
 /* PAGE BREAK */
 .bn-block-content[data-content-type="pageBreak"] > div {
   width: 100%;
-  height: 8px;
-  background: linear-gradient(
-          to right,
-          rgb(125, 121, 122) 50%,
-          rgb(125, 121, 122, 0) 0%
-  );
-  background-position: center;
-  background-size: 16px 1.5px;
-  background-repeat: repeat-x;
+  height: 0;
+  border-top: dotted rgb(125, 121, 122) 4px;
 }
 
 @media print {
