+++ packages/react/src/editor/styles.css
@@ -238,3 +238,7 @@
 .bn-side-menu[data-url="false"] {
   height: 54px;
 }
+
+.bn-side-menu[data-block-type="pageBreak"] {
+  transform: translateY(-10px);
+}
